requirements.txt
pandas==2.2.*
pyarrow>=14.0.0
kss>=4.5.4
PyPDF2>=3.0.1
scikit-learn>=1.4
matplotlib>=3.8
networkx>=3.2
PyYAML>=6.0
regex>=2024.11.6
